auto-analytics-adder - separate general api methods from worker
